+++ crates/sncast/src/starknet_commands/account/mod.rs
@@ -3,12 +3,13 @@ use crate::starknet_commands::account::delete::Delete;
 use crate::starknet_commands::account::deploy::Deploy;
 use crate::starknet_commands::account::import::Import;
 use crate::starknet_commands::account::list::List;
-use anyhow::{Context, Result, anyhow, bail};
+use anyhow::{Context, Result, bail};
 use camino::Utf8PathBuf;
 use clap::{Args, Subcommand};
 use configuration::resolve_config_file;
 use configuration::{load_config, search_config_upwards_relative_to};
 use serde_json::json;
+use sncast::helpers::account::load_accounts;
 use sncast::helpers::rpc::RpcArgs;
 use sncast::{
     AccountType, chain_id_to_network_name, decode_chain_id, helpers::configuration::CastConfig,
@@ -79,9 +80,7 @@ pub fn write_account_to_accounts_file(
         std::fs::write(accounts_file.clone(), "{}")?;
     }
 
-    let contents = std::fs::read_to_string(accounts_file.clone())?;
-    let mut items: serde_json::Value = serde_json::from_str(&contents)
-        .map_err(|_| anyhow!("Failed to parse accounts file at = {}", accounts_file))?;
+    let mut items = load_accounts(accounts_file)?;
 
     let network_name = chain_id_to_network_name(chain_id);